finished DFS
DFS was very quick as it was very similar to BFS, it just used a stack instead of a queue. I also realised that DFS didn't actually always find the shortest path, infact most of the time it would not find a short path but a long winding path. I think I will implement the A* path finding algorithm next