wall implementation, grid cleanup and x and y coord implementation
Turns out walls were very easy to implement, I just had to change a single if statement, so it wasn't a big task at all. The same was the case for the grid cleanup, just remove all visited and path nodes and it was cleaned up. Finally, I renamed the coordinate system to use x-axis and y-axis rather than rows and columns. Much easier to use x and y in grid structures.